requirements.txt
asgiref==3.6.0
Django==4.1.6
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
